refactor: add Container for Dependency Injection
DI를 위한 컨테이너 구현
