+++ README.md
@@ -33,4 +33,4 @@
 ### 2단계 기능
 - [x] 카카오 지도 SDK 적용
 - [x] 검색창 클릭 시 검색 액티비티 이동
-- [ ] MVVM 적용
+- [x] MVVM 적용
