+++ app/src/main/java/campus/tech/kakao/map/MyApplication.kt
@@ -5,9 +5,11 @@ import campus.tech.kakao.map.di.AppContainer
 import com.kakao.vectormap.KakaoMapSdk
 
 class MyApplication : Application() {
+    lateinit var appContainer: AppContainer
 
     override fun onCreate() {
         KakaoMapSdk.init(this,campus.tech.kakao.map.BuildConfig.KAKAO_API_KEY)
         super.onCreate()
+        appContainer = AppContainer(applicationContext)
     }
 }
